Upgraded to Support Cli v4

+++ generator/router.helper.js
@@ -1,8 +1,13 @@
 const fs = require('fs');
+const path = require('path');
 
 module.exports = (api) => {
 	api.onCreateComplete(() => {
-	  const oldFile = api.resolve('src/router.js');
-	  fs.unlinkSync(oldFile);
+	  const oldFile = api.resolve('src/router/index.js');
+    const myFile = path.resolve(__dirname, '../template/src/router/index.js');
+    if (fs.existsSync(oldFile)) {
+      fs.unlinkSync(oldFile);
+      fs.copyFileSync(myFile, oldFile);
+    }
 	});
 };
